requirements-lee.txt
pillow
numpy
tqdm
timm
easydict
pyyaml == 5.4.1
